Computes ad-hoc core inflation using the subitem exclusion method

Computes ad-hoc core inflation using the subitem exclusion method

core_adhoc(core, data, change, weight, code, date)

Arguments

core
A string specifying the inflation core exclusion method: ex0, ex1, ex2 or ex3.

data
Dataframe object with inflation sub-items data.

change
A string with the column name of the percentage change of the inflation sub-items.

weight
A string with the column name of the weight of the inflation sub-items.

code
A string with the column name of the code of the inflation sub-items as provided in get_ipca().

date
A string with the name of the date column.

Value

Calculated core inflation tibble.

Author

Fernando da Silva <fernando@fortietwo.com>

Examples

if (FALSE) {
library(magrittr)
ipca_ex0 <- get_ipca(period = "all") %>%
core_adhoc("ex0", ., "pct_change", "weight", "code", "date")
}